virtualtest-backend/requirements.txt
# ============================================
# VirtuaTest Backend - Python Paketleri
# ============================================
# Her paket ne işe yarıyor açıklamalı

# ----- WEB FRAMEWORK -----
# FastAPI: Modern, hızlı Python web framework
# Avantajları: Otomatik API dökümantasyonu, tip kontrolü, async destek
fastapi==0.109.0

# Uvicorn: ASGI sunucusu - FastAPI'yi çalıştırır
# [standard]: WebSocket, otomatik reload gibi ekstra özellikler
uvicorn[standard]==0.27.0

# ----- DATABASE -----
# SQLAlchemy: ORM (Object Relational Mapper)
# SQL yazmadan Python nesneleriyle veritabanı işlemi
# Örnek: User(email="x") yerine INSERT INTO users yazmazsın
sqlalchemy==2.0.25

# asyncpg: PostgreSQL için async driver
# FastAPI async çalıştığı için async driver şart
asyncpg==0.29.0

# psycopg2-binary: PostgreSQL bağlantısı için (sync işlemler)
# Alembic migration için gerekli
psycopg2-binary==2.9.9

# Alembic: Database migration aracı
# Tablo yapısı değişince mevcut verileri korur
alembic==1.13.1

# ----- GÜVENLİK -----
# python-jose: JWT token oluşturma/doğrulama
# [cryptography]: Şifreleme algoritmaları
python-jose[cryptography]==3.3.0

# passlib: Şifre hashleme (bcrypt algoritması)
# Şifreler düz metin ASLA saklanmaz!
passlib[bcrypt]==1.7.4

# python-multipart: Dosya yükleme desteği
# Speaking modülünde ses dosyası için
python-multipart==0.0.6

# ----- DOĞRULAMA -----
# Pydantic: Veri doğrulama
# API'ye gelen verinin doğru formatta olduğunu kontrol eder
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# ----- YARDIMCI -----
# python-dotenv: .env dosyasını okur
python-dotenv==1.0.1

# httpx: HTTP client - AI servislerine istek için
httpx==0.26.0

bcrypt==4.0.1

google-generativeai==0.8.3
gTTS==2.5.1